# docs/api/openapi.json
{
  "components": {
    "parameters": {
      "AccountID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "账号 UUID",
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "example": {
          "error": "account not found"
        },
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "S3C_TOKEN 中的 Bearer token。可通过 S3C_TOKEN_SCOPES 为单个 token 限定 readonly / prefixes（桶或桶内键前缀）/ accounts / expiresAt；越权请求返回 403，过期 token 返回 401。未在 S3C_TOKEN_SCOPES 中登记的 token 为全权。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "s3client API",
    "version": "v1.0.0",
    "description": "s3client 是面向 S3 兼容对象存储的多账号 Web 控制台。本文档为 /api/* 端点的 OpenAPI 3.0 契约，所有响应均 JSON（除 /api/health 等纯状态端点）。鉴权：Bearer Token（环境变量 S3C_TOKEN，多值逗号分隔）。"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/accounts/{id}/copy-object": {
      "post": {
        "operationId": "copyObject",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs/a.txt",
                "newBucket": "archive",
                "newKey": "a.txt"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "checksumAlgorithm": {
                    "type": "string",
                    "description": "可选；非空时服务端计算并存储全对象校验和（供 verify-checksum 端到端比对）",
                    "enum": [
                      "CRC64NVME",
                      "SHA256",
                      "CRC32C",
                      "SHA1"
                    ]
                  },
                  "ifMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当**目标**对象当前 ETag 匹配时写入"
                  },
                  "ifNoneMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当**目标**对象不存在时写入",
                    "enum": [
                      "*"
                    ]
                  },
                  "key": {
                    "type": "string"
                  },
                  "newBucket": {
                    "type": "string",
                    "format": "可选；省略=同桶"
                  },
                  "newKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key",
                  "newKey"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "copied": "a.txt",
                  "bucket": "archive"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "copied": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "条件字段非法"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "条件不满足（PreconditionFailed）"
          }
        },
        "summary": "单文件复制（不删源）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/copy-prefix": {
      "post": {
        "operationId": "copyPrefix",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "prefix": "docs/",
                "targetBucket": "archive",
                "targetPrefix": "2026/docs/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "targetBucket": {
                    "type": "string",
                    "format": "可选；省略=同桶"
                  },
                  "targetPrefix": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "prefix",
                  "targetPrefix"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "copied": 12,
                  "failed": 0,
                  "total": 12,
                  "truncated": false
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "copied": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    },
                    "failedKeys": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "lastError": {
                      "type": "string"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "含 copied/failed/lastError"
          }
        },
        "summary": "递归复制前缀（同步流式）",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/accounts/{id}/copy-prefix/async": {
      "post": {
        "operationId": "copyPrefixAsync",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "prefix": "docs/",
                "targetBucket": "archive",
                "targetPrefix": "2026/docs/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "targetBucket": {
                    "type": "string",
                    "format": "可选；省略=同桶"
                  },
                  "targetPrefix": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "prefix",
                  "targetPrefix"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "example": {
                  "jobId": "6f1e2d3c-4b5a-6789-abcd-ef0123456789",
                  "total": 12,
                  "truncated": false
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "jobId"
          }
        },
        "summary": "递归复制前缀（异步）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/delete-prefix": {
      "post": {
        "operationId": "deletePrefix",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "prefix": "docs/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "prefix"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "deleted": 11,
                  "failed": 0,
                  "truncated": false
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    },
                    "lastError": {
                      "type": "string"
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "含 deleted/failed/truncated/lastError"
          }
        },
        "summary": "递归删除前缀（同步流式）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/migrate": {
      "post": {
        "operationId": "migrate",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "sourceAccountId": "1f0c2a44-0b1e-4f5a-9c3d-7e8f9a0b1c2d",
                "sourceBucket": "src-bucket",
                "sourceKeys": [
                  "docs/a.txt"
                ],
                "targetAccountId": "2a1b3c4d-5e6f-7081-92a3-b4c5d6e7f809",
                "targetBucket": "dst-bucket",
                "targetPrefix": "migrated/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "string"
                  },
                  "sourceBucket": {
                    "type": "string"
                  },
                  "sourceKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "targetAccountId": {
                    "type": "string"
                  },
                  "targetBucket": {
                    "type": "string"
                  },
                  "targetPrefix": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "migrated": 1,
                  "failed": 0,
                  "failedKeys": []
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "failed": {
                      "type": "integer"
                    },
                    "failedKeys": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "lastError": {
                      "type": "string"
                    },
                    "migrated": {
                      "type": "integer"
                    }
                  }
                }
              }
            },
            "description": "含 migrated/failed/failedKeys/lastError"
          }
        },
        "summary": "同步迁移（流式）",
        "tags": [
          "migrate"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/",
      "description": "同源（前端 Vite 代理或后端 SPA fallback）"
    }
  ],
...
}